pyproject.toml
[tool.poetry]
name = "skypro_cw4"
version = "0.1.0"
description = ""
authors = ["PyDevJ <gitcloudhab@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
